EAMRegistrationSystem/requirements.txt
asgiref==3.6.0
crispy-bootstrap3==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-filter==23.1
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pygments==2.14.0
pytz==2022.7.1
sqlparse==0.4.3
